refactor: centralize firebase config
