+++ js/firebaseConfig.js
@@ -1,0 +1,18 @@
+import { initializeApp } from 'firebase/app';
+import { getAnalytics } from 'firebase/analytics';
+
+const firebaseConfig = {
+  apiKey: "AIzaSyAPPllpKiFOcjqxnuk2tRvithFYKSzkQAc",
+  authDomain: "smarthubultra.firebaseapp.com",
+  databaseURL: "https://smarthubultra-default-rtdb.firebaseio.com",
+  projectId: "smarthubultra",
+  storageBucket: "smarthubultra.firebasestorage.app",
+  messagingSenderId: "12039705608",
+  appId: "1:12039705608:web:f1a4383b245275eaa26dbd",
+  measurementId: "G-V24P3DHL9M"
+};
+
+const app = initializeApp(firebaseConfig);
+const analytics = getAnalytics(app);
+
+export { app, analytics, firebaseConfig };
